+++ evaluation/__init__.py
@@ -1,0 +1,3 @@
+"""
+Evaluation package for the prediction competition.
+"""
